chore: split file for DynamicBone info
