+++ Editor/APIInternal/ComponentInfos.DynamicBone.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 141e8d70d6004da6ad46df6529ede586
+timeCreated: 1700527632
